insider-attack-scenario-pt/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "insider-attack-scenario-pt"
version = "1.0.0"
description = ""
license = "Apache-2.0"
dependencies = [
    "flwr[simulation]>=1.16.0",
    "flwr-datasets[vision]>=0.5.0",
    "torch==2.5.1",
    "torchvision==0.20.1",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.flwr.app]
publisher = "Adam"

[tool.flwr.app.components]
serverapp = "insider_attack_scenario_pt.server_app:app"
clientapp = "insider_attack_scenario_pt.client_app:app"

[tool.flwr.app.config]
num-server-rounds = 5
fraction-fit = 0.5
local-epochs = 1

[tool.flwr.federations]
default = "local-simulation"

[tool.flwr.federations.local-simulation]
options.num-supernodes = 100
